+++ docs/release-notes.md
@@ -53,3 +53,4 @@ User authentication service API. Generates a JWT token for authenticate user in 
 
 #### Version 1.4.0
 - Refactor App-Client-Headers (only ClientId is necessary for endpoints, except Authenticate)
+- Policy UserAdminOrAuthorizedService added